+++ src/test/kotlin/com/github/themartdev/intellijgleam/parser/LetTest.kt
@@ -1,0 +1,9 @@
+package com.github.themartdev.intellijgleam.parser
+
+class LetTest : GleamParsingTestCase() {
+    override fun getTestDataPath(): String = super.getTestDataPath() + "/let"
+
+    fun testSimple() = assertParsedCorrectly()
+
+    fun testAssert() = assertParsedCorrectly()
+}